requirements.txt
fastapi
uvicorn[standard]
requests
pymongo
python-dotenv
pydantic
